integration: factor out the temporary dir creation
